requirements.txt
torch
pyyaml
pandas
matplotlib
pytest
pytest-timeout
seaborn
